all vars moved form useProfileMenu custom hook to context && deletes useProfileMenu hook
